Start defining commitlog.
