+++ .circleci/config.yml
@@ -32,6 +32,8 @@ jobs:
             source $HOME/.cargo/env
             rustup component add rustfmt-preview
             cargo fmt --all -- --check
+            #cargo install cargo-fix
+            # cargo fix
       - run:
           name: Run 'cargo clippy'
           command: |
