+++ starfish-file/src/commitlog/mod.rs
@@ -1,10 +1,37 @@
 use std::path::PathBuf;
 
+pub mod commitlog;
 pub mod segment;
 pub mod segment_manager;
 
 mod descriptor;
 pub use descriptor::Descriptor;
+
+#[derive(Debug, Error)]
+pub enum Error {
+    #[error(display = "Commitlog has been shut down. Cannot add data")]
+    Closed,
+
+    #[error(display = "IO Error: _1")]
+    IO(std::io::Error),
+
+    #[error(display = "Something else failed: _1")]
+    Other(Box<dyn std::error::Error>),
+}
+
+pub type Result<T> = std::result::Result<T, Error>;
+
+impl From<std::io::Error> for Error {
+    fn from(f: std::io::Error) -> Self {
+        Error::IO(f)
+    }
+}
+
+impl From<Box<dyn std::error::Error>> for Error {
+    fn from(f: Box<dyn std::error::Error>) -> Self {
+        Error::Other(f)
+    }
+}
 
 #[derive(Debug, Clone, PartialEq)]
 pub enum SyncMode {
