+++ starfish-file/src/commitlog/segment.rs
@@ -1,5 +1,6 @@
-use crate::commitlog::segment_manager::Error;
 use crate::commitlog::segment_manager::SegmentManager;
+use crate::commitlog::Error;
+use crate::commitlog::Result;
 use crate::fs::File;
 use crate::shared::Shared;
 use std::rc::Rc;
